Update to latest dependencies
In preparation for the move to Rails 5, update to the latest version of
our dependencies.

I carefully updated outdated dependencies as identified by `bundle
outdated`, running the test suite regularly to see if anything broke.

I removed the pinning of nokogiri to 1.6.1. This resulted in a test
failure in the `Markdowner` class because newer versions of Nokogiri
seem to append a newline to the return value of `inner_html`. I updated
the implementation to `chomp` it out, though it likely wasn't going to
impact anything.