Cache bundler dependencies locally
The cache was generated with `bundle package --all --all-platforms` in
order to cache all dependencies for all platforms that one might run the
app on.

This was done for the live workshop so users that clone the repo will
have all necessary data locally and will not require an internet
connection.