Update dependencies for Rails 5
I first updated the `Gemfile` constraint for `rails` to pick up the beta
(and any subsequent RCs or releases). Then I tried `bundle update rails`
but was informed of a conflict with `exception_notifier`.

I went to the repo for that gem and poked around to find an issue and PR
for Rails 5. I subscribed to both on GitHub so I'll know when they are
merged or if there are problems. In the meantime, I updated the gemfile
dependency to point at the branch for the Rails 5 pr. Then I `bundle
update rails exception_notifier`

Now I get a conflict with various Rails gems and `rspec-rails`. So I
head off to the `rspec-rails` repo on GitHub. I notice that the current
master branch is for release series 3.5 and that I'm on 3.4.2 locally. I
check RubyGems and see that there's a 3.5.0.beta2 release. I update my
Gemfile dependency to pick that up and any future 3.x releases.

Finally `bundle update rails exception_notifier rspec-rails` and we have
successfully bundled with Rails 5. Our tests fail, but that's not a
conern right now. COMMIT!