Fix failing EmailParser tests
All of the EmailParser tests were failing. From investigating the
failures, it looked like the email parser was setting the email body to
nil, regardless of what was passed in.

To assist in debugging, I added the `pry-rails` gem, which lets us get
into the running process and poke around with `binding.pry`.

I noticed that the code in EmailParser that reads the provided text into
a Mail object had a `rescue` statement that was swallowing all errors
without logging. I put a `binding.pry` statement in that `rescue` and
re-ran the test to see if it was an exception causing the email to be
`nil`. Indeed it was: `NoMethodError` for `Utils.silence_stream`.

`Utils` is in `app/extras`. Indeed, the `silence_stream` method is an
instance method. Making it a class method fixes the error, but shouldn't
this have failed on Rails 4.2 as well? Surely, Rails is not somehow
making instance methods into class methods on plain-old Ruby objects,
right?

Turns out, Rails 4.2 and earlier monkey patched a `silence_stream`
method onto `Kernel` (which is where common methods such as `puts` are
defined). While the code appears to be somehow managing to call the
`silence_stream` instance method on `Utils`, it's actually calling the
method that Rails monkey patched onto Kernel. The method does the same
thing as the one defined in `Utils` so the author did not notice this
issue.

It's perhaps worth noting that test-driving this code when it was
originally written would have caught this error sooner. Running a test
that you expected to fail due to an un-implemented `silence_stream`
method would have actually passed.

Finally, with the test now working I swapped out my `binding.pry` with a
logger statement so at least some evidence of this exception would
appear in logs.